Outro commit

+++ src/app/manter-contatos/manter-contatos.component.ts
@@ -24,12 +24,20 @@ export class ManterContatosComponent implements OnInit {
     this.contatoService.post(this.contato).subscribe(resultado => {
       this.contato = { id: null, tipo: "", telefone: "", usuario: { id_usu: null, nome: "", email: "", cpf: "", dt_nascimento: "", sexo: "", senha: "" }};
     });
+    alert("Contato salvo com sucesso!");
   }
 
   excluir(id) {
-    this.contatoService.delete(this.contato.id).subscribe(resultado => {
-      this.contato = { id: null, tipo: "", telefone: "", usuario: { id_usu: null, nome: "", email: "", cpf: "", dt_nascimento: "", sexo: "", senha: "" }};
-    });
+    var r = confirm("Você realmente deseja remover esse contato?");
+
+    if (r == true) {
+      this.contatoService.delete(this.contato.id).subscribe(resultado => {
+        this.contato = { id: null, tipo: "", telefone: "", usuario: { id_usu: null, nome: "", email: "", cpf: "", dt_nascimento: "", sexo: "", senha: "" }};
+      });
+      alert("Contato removido com sucesso!");
+    } else {
+      alert("Contato não foi removido!");
+    }
   }
 
   consultar(id) {
@@ -41,6 +49,27 @@ export class ManterContatosComponent implements OnInit {
         usuario: dados.usuario
       };
     });
+  }
+
+  consultarUsuario(cpf) {
+    cpf = cpf.replace(/\D/g, '');
+
+    if (cpf != null && cpf !== '') {
+      this.contatoService.getByCpf(cpf).subscribe(dados => {
+        this.contato.usuario = {
+          id: dados.id,
+          nome: dados.nome,
+          email: dados.email,
+          cpf: dados.cpf,
+          dt_nascimento: dados.dt_nascimento,
+          sexo: dados.sexo,
+          senha: dados.senha
+        };
+      });
+    } else {
+      this.contato.usuario = { id: null, nome: "", email: "", cpf: "", dt_nascimento: "", sexo: "", senha: "" };
+    }
+    console.log(this.contato.usuario);
   }
 
   limpar(form) {
