+++ src/app/manter-rotas/manter-rotas.component.ts
@@ -8,11 +8,13 @@ import { RotaServiceService } from '../manter-rotas/rota-service.service';
 })
 export class ManterRotasComponent implements OnInit {
 
-  rota: { id, data, horario, inicio, fim, status, verificador, veiculo, contribuicao} = 
-  { id: null, data:"", horario:"", inicio:"", fim:"", status:"", verificador:"", 
-  veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null,  usuario: null } , 
-  contribuicao: { id: null, tipo: "", valor: ""}};
-  
+  rota: { id, data, horario, inicio, fim, status, verificador, veiculo, contribuicao } =
+    {
+      id: null, data: "", horario: "", inicio: "", fim: "", status: "", verificador: "",
+      veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null, usuario: null },
+      contribuicao: { id: null, tipo: "", valor: "" }
+    };
+
   rotas;
   veiculos;
   contribuicoes;
@@ -20,26 +22,36 @@ export class ManterRotasComponent implements OnInit {
   constructor(private rotaService: RotaServiceService) { }
 
   ngOnInit(): void {
-    this.rotaService.getContribuicoes().subscribe(resultado => {this.contribuicoes = resultado});
-    this.rotaService.getVeiculos().subscribe(resultado => {this.veiculos = resultado});
+    this.rotaService.getContribuicoes().subscribe(resultado => { this.contribuicoes = resultado });
+    this.rotaService.getVeiculos().subscribe(resultado => { this.veiculos = resultado });
   }
 
   salvar() {
     console.log(this.rota);
     this.rotaService.post(this.rota).subscribe(resultado => {
-      this.rota = { id: null, data:"", horario:"", inicio:"", fim:"", status:"", verificador:"", 
-      veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null,  usuario: null } , 
-      contribuicao: { id: null, tipo: "", valor: ""} };
+      this.rota = {
+        id: null, data: "", horario: "", inicio: "", fim: "", status: "", verificador: "",
+        veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null, usuario: null },
+        contribuicao: { id: null, tipo: "", valor: "" }
+      };
     });
   }
 
   excluir(id) {
-    this.rotaService.delete(this.rota.id).subscribe(resultado => {
-      this.rota = { id: null, data:"", horario:"", inicio:"", fim:"", status:"", verificador:"", 
-      veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null,  usuario: null } , 
-      contribuicao: { id: null, tipo: "", valor: ""} };
+    var r = confirm("Você realmente deseja remover esse usuário?");
 
-    });
+    if (r == true) {
+      this.rotaService.delete(this.rota.id).subscribe(resultado => {
+        this.rota = {
+          id: null, data: "", horario: "", inicio: "", fim: "", status: "", verificador: "",
+          veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null, usuario: null },
+          contribuicao: { id: null, tipo: "", valor: "" }
+        };
+      });
+      alert("Usuário removido com sucesso!");
+    } else {
+      alert("Usuário não foi removido!");
+    }
   }
 
   consultar(verificador) {
@@ -58,22 +70,54 @@ export class ManterRotasComponent implements OnInit {
         };
       });
     } else {
-      this.rota = { id: null, data:"", horario:"", inicio:"", fim:"", status:"", verificador:"", 
-      veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null,  usuario: null } , 
-      contribuicao: { id: null, tipo: "", valor: ""} };
+      this.rota = {
+        id: null, data: "", horario: "", inicio: "", fim: "", status: "", verificador: "",
+        veiculo: { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: null, tipo: "", capacidade: null, usuario: null },
+        contribuicao: { id: null, tipo: "", valor: "" }
+      };
     }
+  }
+
+  consultarVeiculo(placa) {
+    if (placa != null && placa !== '') {
+      this.rotaService.getByPlaca(placa).subscribe(dados => {
+        this.rota.veiculo = {
+          id: dados.id,
+          placa: dados.placa,
+          renavam: dados.renavam,
+          modelo: dados.modelo,
+          cor: dados.cor,
+          ano_fabricacao: dados.ano_fabricacao,
+          tipo: dados.tipo,
+          capacidade: dados.capacidade,
+          usuario: dados.usuario
+        };
+      });
+    } else {
+      this.rota.veiculo = { id: null, placa: "", renavam: "", modelo: "", cor: "", ano_fabricacao: "", tipo: "", capacidade: "", usuario: { id: null, nome: "", email: "", cpf: "", dt_nascimento: "", sexo: "", senha: "" } };
+    }
+  }
+
+  consultarContribuicao(id) {
+    this.rotaService.getByIdContribuicoes(id).subscribe(dados => {
+      this.rota.contribuicao = {
+        id: dados.id,
+        tipo: dados.tipo,
+        valor: dados.valor
+      };
+    });
   }
 
   limpar(form) {
     form.reset();
   }
 
-  gerarVerificador(tamanho){
+  gerarVerificador(tamanho) {
     let cod = '';
 
-    do{
+    do {
       cod += Math.random().toString(36).substr(2);
-    } while(cod.length < tamanho);
+    } while (cod.length < tamanho);
 
     cod = cod.substr(0, tamanho);
     cod.toLowerCase();
